+++ analyzer-service/calculations.py
@@ -3,6 +3,7 @@ import pandas as pd
 from scipy.stats import norm
 import statsmodels.api as sm
 from datetime import datetime
+from sklearn.svm import SVC
 
 import time
 
@@ -110,3 +111,126 @@ def calculateSharpeRatio(symbolER, symbolSD):
   rf = 0.025
   sharpe = (symbolER - rf) / symbolSD
   return round(sharpe,2)
+
+####################################################################
+## Investability Index Functions
+'''
+Algorithm
+---------
+Step 1: For all the tickers pull the last 12 months of data split into 6 month chunks
+Step 2: For Each: did the ticker beat the SP500 in the second chunk? (label)
+Step 3: For Each: Calculate the 6 calcs for the first 6 month chunk
+Step 4: Train Classification model
+Step 5: For Each: Test on second chunk with calc values from the db
+Step 6: For Each: Output probability
+Step 7: For Each: Write value to db
+'''
+
+def modifiedFilterData(symbolData):
+  # modified to get 6 month chunk from 1 year ago to 6 months ago
+  df = pd.DataFrame.from_dict(symbolData)
+  df = df.T
+  df.reset_index(level=0, inplace=True)
+
+  current_date = datetime.today()
+  sixMonthDate = (current_date - pd.Timedelta(weeks=26)) - pd.Timedelta(days=1)
+  yearDate = (current_date - pd.Timedelta(days=365)) - pd.Timedelta(days=1)
+
+  df = df.loc[0:300,:]
+  for i in range(len(df.loc[:, 'index'])):
+    df.loc[i,'index'] = datetime.strptime(df.loc[i,'index'] , '%Y-%m-%d')
+
+  resultsDF = df.loc[(df['index'] > yearDate) & (df['index'] < sixMonthDate)]
+  return resultsDF
+
+def trainInvestabilityIndexModel():
+  # Step 1: Pull data from load historical
+  symbols = pd.read_csv('DOW30.csv')
+  symbols = symbols.iloc[:,0].values.tolist()
+
+  # Get data for recent 6 month chunk
+  indexData = loadHistorical('SPY')
+  indexDataRecent = filterData(indexData)
+  indexDataRecent = indexDataRecent.iloc[::-1]
+
+  # Get SPY return over last 6 months
+  r_spy = sixMonthReturn(indexDataRecent)
+
+  # Get data for old 6 mon chunk (used in calcs)
+  indexDataOld = modifiedFilterData(indexData)
+  indexDataOld = indexDataOld.iloc[::-1]
+  bReturnsTrain = getPctReturns(indexDataOld)
+
+  ## TEST:
+  symbols = ['MU', 'AAPL', 'MSFT']
+  ## END TEST
+
+  datasetTrain = pd.DataFrame(columns=['Symbol','VaR', 'B', 'SD', 'r2', 'ER', 'SR', 'isBeat'])
+  # Step 2: Get SPY+tickers return for next 6 month period, label if beat or not
+  # Step 3: Call the analysis handler for train calcs + create dataframe
+  for ticker in symbols:
+    tickerData = loadHistorical(ticker)
+    tickerData = modifiedFilterData(tickerData)
+    tickerData = tickerData.iloc[::-1]
+    pReturnsTrain = getPctReturns(tickerData)
+    r_symbol = sixMonthReturn(tickerData)
+
+    valueAtRisk = calculateValueAtRisk(pReturnsTrain)
+    valueAtRisk = valueAtRisk["Value at Risk 95%"]
+    beta = calculateBeta(pReturnsTrain, bReturnsTrain)
+    standardDeviation = calculateStandardDeviation(pReturnsTrain)
+    rSquared = calculateRSquared(pReturnsTrain, bReturnsTrain)
+    expectedReturn = calculateExpectedReturn(indexDataOld, beta)
+    sharpeRatio = calculateSharpeRatio(expectedReturn, standardDeviation)
+    isBeat = r_symbol > r_spy
+
+    insertList = [ticker, valueAtRisk, beta, standardDeviation, rSquared, expectedReturn, sharpeRatio, isBeat]
+    datasetTrain.loc[len(datasetTrain)] = insertList
+
+  # Machine Learning Preprocessing
+  X_train = datasetTrain.iloc[:, 1:7]
+  y_train = datasetTrain.iloc[:, 7]
+  y_train=y_train.astype('bool')
+
+  # Feature Scaling
+  # from sklearn.preprocessing import StandardScaler
+  # sc = StandardScaler()
+  # X_train = sc.fit_transform(X_train)
+  # X_test = sc.transform(X_test)
+
+  # Step 4: Train the SVC on the dataframe w/ labels
+  # Fitting classifier to the Training set
+  prob_classifier = SVC(kernel = 'rbf', probability=True)
+  prob_classifier.fit(X_train, y_train,)
+
+  # Return trained model
+  return prob_classifier
+
+
+def computeInvestabilityIndex(trainedModel):
+  # Build test data
+  datasetTest = pd.DataFrame(columns=['Symbol','VaR', 'B', 'SD', 'r2', 'ER', 'SR'])
+  symbol = "MSFT"
+  collection = app.AnalysisCollection
+  result = collection.find_one({"symbol": "{}".format(symbol)})
+
+  valueAtRisk = result['valueAtRisk']
+  beta = result['beta']
+  standardDeviation = result['standardDeviation']
+  rSquared = result['rSquared']
+  expectedReturn = result['expectedReturn']
+  sharpeRatio = result['sharpeRatio']
+
+  insertList = [symbol, valueAtRisk, beta, standardDeviation, rSquared, expectedReturn, sharpeRatio]
+  datasetTest.loc[len(datasetTest)] = insertList
+
+  X_test = datasetTest.iloc[:, 1:7]
+  isBeat = trainedModel.predict_proba(X_test)
+  return str(isBeat[0])
+
+
+  # Step 5: Run predict on each ticker with 6 calcs from db
+  # Step 6: Output probability
+  # Step 7: Truncate and write probability to the corresponding collection
+  # Return status complete
+  return 0
